# avoiding compiler optimizations