@benoit-bst 2017-10-02
Thanks David for paying attention to the results. By looking at the ASM code it becomes obvious. The naive implementation of Atoi is therefore the fastest.